DataViews Extensibility: Allow unregistering restore post action
